remove direct dependency on Test::Mojibake
This plugin doesn't use Test::Mojibake, but generates a test that uses
it. Building a dist using this plugin doesn't need Test::Mojibake for
anything, and is an extraneous dependency if you don't intent to run the
test, or are running it on a different system.

Unicode::CheckUTF8 is only indirectly used via Test::Mojibake, so it can
be removed as well.